backend/requirements/base.txt
Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
cohere>=1.0.0
groq>=0.1.0
aiohttp>=3.9.1 
# asyncio>=3.4.3
playwright>=1.44.1
psycopg2-binary==2.9.9
pandas>=2.1.1
fuzzywuzzy>=0.15.0
xlsxwriter>=3.1.0
openpyxl>=3.1.2
python-levenshtein>=0.21.0
scikit-learn>=1.4.0
numpy>=1.26.0
# For error logging system
django-filter>=23.5
django-extensions>=3.2.3
channels>=4.0.0
# Azure Blob Storage
azure-storage-blob>=12.18.0
celery>=5.3.0
# Amazon Ads Reports dependencies

# API Clients
requests>=2.28.0

# Utilities
python-dateutil>=2.8.0 

# Firebase
firebase-admin>=6.0.0
pyrebase4>=4.6.0 